♻(ingestion-presentation) isolate metiers views
