+++ src/web/presentation/ingestion/views.py
@@ -11,7 +11,6 @@ from drf_spectacular.utils import (
 from pydantic import ValidationError
 from rest_framework import serializers, status
 from rest_framework import serializers as drf_serializers
-from rest_framework.exceptions import ValidationError as DRFValidationError
 from rest_framework.parsers import FormParser, JSONParser, MultiPartParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -20,7 +19,6 @@ from rest_framework_simplejwt.authentication import JWTAuthentication
 from application.ingestion.interfaces.archive_offer_by_reference_input import (
     ArchiveOfferByReferenceInput,
 )
-from application.ingestion.interfaces.list_metiers_input import GetFilteredMetiersInput
 from application.ingestion.interfaces.list_offers_input import GetFilteredOffersInput
 from application.ingestion.interfaces.load_documents_input import LoadDocumentsInput
 from application.ingestion.interfaces.load_operation_type import LoadOperationType
@@ -38,8 +36,6 @@ from presentation.ingestion.openapi import (
     ARCHIVE_OFFER_EXAMPLES,
     CONCOURS_UPLOAD_DESCRIPTION,
     CONCOURS_UPLOAD_EXAMPLES,
-    LIST_METIERS_DESCRIPTION,
-    LIST_METIERS_EXAMPLES,
     LIST_OFFERS_DESCRIPTION,
     LIST_OFFERS_EXAMPLES,
     UPSERT_OFFERS_DESCRIPTION,
@@ -52,8 +48,6 @@ from presentation.ingestion.serializers import (
     ConcoursUploadResponseSerializer,
     GenericErrorSerializer,
     IdentityInputSerializer,
-    ListMetiersFiltersSerializer,
-    ListMetiersResponseSerializer,
     ListOffersFiltersSerializer,
     ListOffersResponseSerializer,
     NoValidRowsErrorSerializer,
@@ -338,57 +332,6 @@ class ConcoursUploadView(APIView):
             },
             status=status.HTTP_201_CREATED,
         )
-
-
-@extend_schema(
-    summary="Liste des métiers",
-    description=LIST_METIERS_DESCRIPTION,
-    examples=LIST_METIERS_EXAMPLES,
-    tags=["metiers"],
-    parameters=[ListMetiersFiltersSerializer],
-    responses={
-        200: ListMetiersResponseSerializer,
-        400: GenericErrorSerializer,
-        401: TokenErrorSerializer,
-        500: GenericErrorSerializer,
-    },
-)
-class MetiersListView(APIView):
-    serializer_class = ListMetiersResponseSerializer
-    usecase = None
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.container = create_ingestion_container()
-        self.logger = self.container.logger_service()
-        if self.usecase is None:
-            self.usecase = self.container.list_metiers_usecase()
-
-    def get(self, request):
-        try:
-            filters = ListMetiersFiltersSerializer(data=self.request.query_params)
-            filters.is_valid(raise_exception=True)
-            input_data = GetFilteredMetiersInput(**filters.validated_data)
-
-            result = self.usecase.execute(input_data)
-
-            paginator = IngestionPagination()
-            items = paginator.paginate(result, request)
-            return paginator.get_paginated_response(
-                ListMetiersResponseSerializer(items, many=True).data
-            )
-        except DRFValidationError as e:
-            serializer = GenericErrorSerializer({"error": str(e)})
-            return Response(
-                serializer.data,
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-        except Exception as e:
-            self.logger.error("Unexpected error in MetiersListView: %s", str(e))
-            serializer = GenericErrorSerializer({"error": "Unexpected error"})
-            return Response(
-                serializer.data, status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
 
 
 @extend_schema(
